remove and/or poop
