+++ service/tbc/crawler.go
@@ -9,7 +9,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"runtime"
 	"sync"
 	"time"
 
@@ -18,7 +17,6 @@ import (
 	"github.com/btcsuite/btcd/chaincfg/chainhash"
 	"github.com/btcsuite/btcd/txscript"
 	"github.com/btcsuite/btcd/wire"
-	"github.com/dustin/go-humanize"
 
 	"github.com/hemilabs/heminetwork/v2/api/tbcapi"
 	"github.com/hemilabs/heminetwork/v2/database"
@@ -495,22 +493,6 @@ func (s *Server) headerAndBlock(ctx context.Context, hash chainhash.Hash) (*tbcd
 	return bh, b, nil
 }
 
-func logMemStats() {
-	var mem runtime.MemStats
-	runtime.ReadMemStats(&mem)
-
-	// Go memory statistics are hard to interpret but the following list is
-	// an approximation:
-	//	Alloc is currently allocated memory
-	// 	TotalAlloc is all memory allocated over time
-	// 	Sys is basically a peak memory use
-	log.Infof("Alloc = %v, TotalAlloc = %v, Sys = %v, NumGC = %v\n",
-		humanize.IBytes(mem.Alloc),
-		humanize.IBytes(mem.TotalAlloc),
-		humanize.IBytes(mem.Sys),
-		mem.NumGC)
-}
-
 func processUtxos(block *btcutil.Block, direction int, utxos map[tbcd.Outpoint]tbcd.CacheOutput) error {
 	txs := block.Transactions()
 	for _, tx := range txs {
@@ -773,327 +755,6 @@ func (s *Server) fixupCacheChannel(ctx context.Context, b *btcutil.Block, utxos 
 	return nil
 }
 
-//// indexUtxosInBlocks indexes utxos from the last processed block until the
-//// provided end hash, inclusive. It returns the number of blocks processed and
-//// the last hash it has processed.
-//func (s *Server) indexUtxosInBlocks(ctx context.Context, endHash *chainhash.Hash, utxos map[tbcd.Outpoint]tbcd.CacheOutput) (int, *HashHeight, error) {
-//	log.Tracef("indexUtxoBlocks")
-//	defer log.Tracef("indexUtxoBlocks exit")
-//
-//	// indicates if we have processed endHash and thus have hit the exit
-//	// condition.
-//	var last *HashHeight
-//
-//	// Find start hash
-//	utxoHH, err := s.UtxoIndexHash(ctx)
-//	if err != nil {
-//		return 0, last, fmt.Errorf("utxo index hash: %w", err)
-//	}
-//
-//	// If we have a real block move forward to the next block since we
-//	// already indexed the last block.
-//	hh := utxoHH
-//	if !hh.Hash.IsEqual(s.chainParams.GenesisHash) {
-//		hh, err = s.nextCanonicalBlockheader(ctx, endHash, hh)
-//		if err != nil {
-//			return 0, last, fmt.Errorf("utxo next block %v: %w", hh, err)
-//		}
-//	}
-//
-//	utxosPercentage := 95 // flush cache at >95% capacity
-//	blocksProcessed := 0
-//	for {
-//		log.Debugf("indexing utxos: %v", hh)
-//
-//		bh, b, err := s.headerAndBlock(ctx, hh.Hash)
-//		if err != nil {
-//			return 0, last, err
-//		}
-//
-//		err = s.fixupCache(ctx, b, utxos)
-//		if err != nil {
-//			return 0, last, fmt.Errorf("process utxos fixup %v: %w", hh, err)
-//		}
-//		err = processUtxos(b, 1, utxos)
-//		if err != nil {
-//			return 0, last, fmt.Errorf("process utxos %v: %w", hh, err)
-//		}
-//
-//		blocksProcessed++
-//
-//		// Try not to overshoot the cache to prevent costly allocations
-//		cp := len(utxos) * 100 / s.cfg.MaxCachedTxs
-//		if bh.Height%10000 == 0 || cp > utxosPercentage || blocksProcessed == 1 {
-//			log.Infof("Utxo indexer: %v utxo cache %v%%", hh, cp)
-//		}
-//		if cp > utxosPercentage {
-//			// Set utxosMax to the largest utxo capacity seen
-//			s.cfg.MaxCachedTxs = max(len(utxos), s.cfg.MaxCachedTxs)
-//			last = hh
-//			// Flush
-//			break
-//		}
-//
-//		// Exit if we processed the provided end hash
-//		if endHash.IsEqual(&hh.Hash) {
-//			last = hh
-//			break
-//		}
-//
-//		// Move to next block
-//		hh, err = s.nextCanonicalBlockheader(ctx, endHash, hh)
-//		if err != nil {
-//			return 0, last, fmt.Errorf("utxo next block %v: %w", hh, err)
-//		}
-//	}
-//
-//	return blocksProcessed, last, nil
-//}
-//
-//// unindexUtxosInBlocks unindexes utxos from the last processed block until the
-//// provided end hash, inclusive. It returns the number of blocks processed and
-//// the last hash it has processed.
-//func (s *Server) unindexUtxosInBlocks(ctx context.Context, endHash *chainhash.Hash, utxos map[tbcd.Outpoint]tbcd.CacheOutput) (int, *HashHeight, error) {
-//	log.Tracef("unindexUtxoBlocks")
-//	defer log.Tracef("unindexUtxoBlocks exit")
-//
-//	// indicates if we have processed endHash and thus have hit the exit
-//	// condition.
-//	var last *HashHeight
-//
-//	// Find start hash
-//	utxoHH, err := s.UtxoIndexHash(ctx)
-//	if err != nil {
-//		return 0, last, fmt.Errorf("utxo index hash: %w", err)
-//	}
-//
-//	utxosPercentage := 95 // flush cache at >95% capacity
-//	blocksProcessed := 0
-//	hh := utxoHH
-//	for {
-//		log.Debugf("unindexing utxos: %v", hh)
-//
-//		hash := hh.Hash
-//		bh, err := s.db.BlockHeaderByHash(ctx, hash)
-//		if err != nil {
-//			return 0, last, fmt.Errorf("block header %v: %w", hash, err)
-//		}
-//
-//		// Exit if we processed the provided end hash
-//		if endHash.IsEqual(&hash) {
-//			last = hh
-//			break
-//		}
-//
-//		// Index block
-//		b, err := s.db.BlockByHash(ctx, bh.Hash)
-//		if err != nil {
-//			return 0, last, fmt.Errorf("block by hash %v: %w", bh, err)
-//		}
-//
-//		err = s.unprocessUtxos(ctx, b, utxos)
-//		if err != nil {
-//			return 0, last, fmt.Errorf("process utxos %v: %w", hh, err)
-//		}
-//
-//		// Add tx's back to the mempool.
-//		// if s.cfg.MempoolEnabled {
-//		//	// XXX this may not be the right spot.
-//		//	txHashes, _ := b.MsgBlock().TxHashes()
-//		//	_ = s.mempool.txsRemove(ctx, txHashes)
-//		// }
-//
-//		blocksProcessed++
-//
-//		// Try not to overshoot the cache to prevent costly allocations
-//		cp := len(utxos) * 100 / s.cfg.MaxCachedTxs
-//		if bh.Height%10000 == 0 || cp > utxosPercentage || blocksProcessed == 1 {
-//			log.Infof("UTxo unindexer: %v utxo cache %v%%", hh, cp)
-//		}
-//
-//		// Move to previous block
-//		height := bh.Height - 1
-//		pbh, err := s.db.BlockHeaderByHash(ctx, *bh.ParentHash())
-//		if err != nil {
-//			return 0, last, fmt.Errorf("block headers by height %v: %w",
-//				height, err)
-//		}
-//		hh.Hash = *pbh.BlockHash()
-//		hh.Height = pbh.Height
-//
-//		// We check overflow AFTER obtaining the previous hash so that
-//		// we can update the database with the LAST processed block.
-//		if cp > utxosPercentage {
-//			// Set txsMax to the largest tx capacity seen
-//			s.cfg.MaxCachedTxs = max(len(utxos), s.cfg.MaxCachedTxs)
-//			last = hh
-//			// Flush
-//			break
-//		}
-//	}
-//
-//	return blocksProcessed, last, nil
-//}
-//
-//func (s *Server) utxoIndexerUnwind(ctx context.Context, startBH, endBH *tbcd.BlockHeader) error {
-//	log.Tracef("utxoIndexerUnwind")
-//	defer log.Tracef("utxoIndexerUnwind exit")
-//
-//	// XXX dedup with TxIndexedWind; it's basically the same code but with the direction, start and endhas flipped
-//	s.mtx.Lock()
-//	if !s.indexing {
-//		// XXX this prob should be an error but pusnish bad callers for now
-//		s.mtx.Unlock()
-//		panic("utxoIndexerUnwind not true")
-//	}
-//	s.mtx.Unlock()
-//
-//	// Allocate here so that we don't waste space when not indexing.
-//	utxos := make(map[tbcd.Outpoint]tbcd.CacheOutput, s.cfg.MaxCachedTxs)
-//	defer clear(utxos)
-//
-//	log.Infof("Start unwinding UTxos at hash %v height %v", startBH, startBH.Height)
-//	log.Infof("End unwinding UTxos at hash %v height %v", endBH, endBH.Height)
-//	endHash := endBH.BlockHash()
-//	for {
-//		start := time.Now()
-//		blocksProcessed, last, err := s.unindexUtxosInBlocks(ctx, endHash, utxos)
-//		if err != nil {
-//			return fmt.Errorf("unindex utxos in blocks: %w", err)
-//		}
-//		if blocksProcessed == 0 {
-//			return nil
-//		}
-//		utxosCached := len(utxos)
-//		log.Infof("UTxo unwinder blocks processed %v in %v transactions cached %v cache unused %v avg tx/blk %v",
-//			blocksProcessed, time.Since(start), utxosCached,
-//			s.cfg.MaxCachedTxs-utxosCached, utxosCached/blocksProcessed)
-//
-//		// Flush to disk
-//		start = time.Now()
-//		if err = s.db.BlockUtxoUpdate(ctx, -1, utxos, last.Hash); err != nil {
-//			return fmt.Errorf("block utxo update: %w", err)
-//		}
-//		// leveldb does all kinds of allocations, force GC to lower
-//		// memory pressure.
-//		logMemStats()
-//		runtime.GC()
-//
-//		log.Infof("Flushing unwind utxos complete %v took %v",
-//			utxosCached, time.Since(start))
-//
-//		if endHash.IsEqual(&last.Hash) {
-//			break
-//		}
-//	}
-//
-//	return nil
-//}
-//
-//func (s *Server) utxoIndexerWind(ctx context.Context, startBH, endBH *tbcd.BlockHeader) error {
-//	log.Tracef("utxoIndexerWind")
-//	defer log.Tracef("utxoIndexerWind exit")
-//
-//	s.mtx.Lock()
-//	if !s.indexing {
-//		// XXX this prob should be an error but pusnish bad callers for now
-//		s.mtx.Unlock()
-//		panic("utxoIndexerWind not true")
-//	}
-//	s.mtx.Unlock()
-//
-//	// Allocate here so that we don't waste space when not indexing.
-//	utxos := make(map[tbcd.Outpoint]tbcd.CacheOutput, s.cfg.MaxCachedTxs)
-//	defer clear(utxos)
-//
-//	log.Infof("Start indexing UTxos at hash %v height %v", startBH, startBH.Height)
-//	log.Infof("End indexing UTxos at hash %v height %v", endBH, endBH.Height)
-//	endHash := endBH.BlockHash()
-//	for {
-//		start := time.Now()
-//		blocksProcessed, last, err := s.indexUtxosInBlocks(ctx, endHash, utxos)
-//		if err != nil {
-//			return fmt.Errorf("index blocks: %w", err)
-//		}
-//		if blocksProcessed == 0 {
-//			return nil
-//		}
-//		utxosCached := len(utxos)
-//		log.Infof("Utxo indexer blocks processed %v in %v utxos cached %v cache unused %v avg tx/blk %v",
-//			blocksProcessed, time.Since(start), utxosCached,
-//			s.cfg.MaxCachedTxs-utxosCached, utxosCached/blocksProcessed)
-//
-//		// Flush to disk
-//		start = time.Now()
-//		if err = s.db.BlockUtxoUpdate(ctx, 1, utxos, last.Hash); err != nil {
-//			return fmt.Errorf("block tx update: %w", err)
-//		}
-//
-//		// leveldb does all kinds of allocations, force GC to lower
-//		// memory pressure.
-//		logMemStats()
-//		runtime.GC()
-//
-//		log.Infof("Flushing utxos complete %v took %v",
-//			utxosCached, time.Since(start))
-//
-//		if endHash.IsEqual(&last.Hash) {
-//			break
-//		}
-//	}
-//
-//	return nil
-//}
-//
-//func (s *Server) utxoIndexer(ctx context.Context, endHash chainhash.Hash) error {
-//	log.Tracef("utxoIndexer")
-//	defer log.Tracef("utxoIndexer exit")
-//
-//	s.mtx.Lock()
-//	if !s.indexing {
-//		// XXX this prob should be an error but pusnish bad callers for now
-//		s.mtx.Unlock()
-//		panic("utxoIndexer indexing not true")
-//	}
-//	s.mtx.Unlock()
-//	// XXX this is basically duplicate from UtxoIndexIsLinear
-//
-//	// Verify exit condition hash
-//	endBH, err := s.db.BlockHeaderByHash(ctx, endHash)
-//	if err != nil {
-//		return fmt.Errorf("blockheader hash: %w", err)
-//	}
-//
-//	// Verify start point is not after the end point
-//	utxoHH, err := s.UtxoIndexHash(ctx)
-//	if err != nil {
-//		return fmt.Errorf("utxo index hash: %w", err)
-//	}
-//
-//	// XXX make sure there is no gap between start and end or vice versa.
-//	startBH, err := s.db.BlockHeaderByHash(ctx, utxoHH.Hash)
-//	if err != nil {
-//		return fmt.Errorf("blockheader hash: %w", err)
-//	}
-//	direction, err := s.UtxoIndexIsLinear(ctx, endHash)
-//	if err != nil {
-//		return fmt.Errorf("utxo index is linear: %w", err)
-//	}
-//	log.Debugf("startbh %v", startBH.HH())
-//	log.Debugf("endHash %v", endHash)
-//	log.Debugf("direction %v", direction)
-//	switch direction {
-//	case 1:
-//		return s.utxoIndexerWind(ctx, startBH, endBH)
-//	case -1:
-//		return s.utxoIndexerUnwind(ctx, startBH, endBH)
-//	case 0:
-//		// Because we call UtxoIndexIsLinear we know it's the same block.
-//		return nil
-//	}
-//	return fmt.Errorf("invalid direction: %v", direction)
-//}
-
 func processTxs(block *btcutil.Block, direction int, txsCache map[tbcd.TxKey]*tbcd.TxValue) error {
 	blockHash := block.Hash()
 	txs := block.Transactions()
@@ -1119,324 +780,6 @@ func processTxs(block *btcutil.Block, direction int, txsCache map[tbcd.TxKey]*tb
 	}
 	return nil
 }
-
-//// indexTxsInBlocks indexes txs from the last processed block until the
-//// provided end hash, inclusive. It returns the number of blocks processed and
-//// the last hash it has processed.
-//func (s *Server) indexTxsInBlocks(ctx context.Context, endHash *chainhash.Hash, txs map[tbcd.TxKey]*tbcd.TxValue) (int, *HashHeight, error) {
-//	log.Tracef("indexTxsInBlocks")
-//	defer log.Tracef("indexTxsInBlocks exit")
-//
-//	// indicates if we have processed endHash and thus have hit the exit
-//	// condition.
-//	var last *HashHeight
-//
-//	// Find start hash
-//	txHH, err := s.TxIndexHash(ctx)
-//	if err != nil {
-//		return 0, last, fmt.Errorf("tx index hash: %w", err)
-//	}
-//
-//	// If we have a real block move forward to the next block since we
-//	// already indexed the last block.
-//	hh := txHH
-//	if !hh.Hash.IsEqual(s.chainParams.GenesisHash) {
-//		hh, err = s.nextCanonicalBlockheader(ctx, endHash, hh)
-//		if err != nil {
-//			return 0, last, fmt.Errorf("tx next block %v: %w", hh, err)
-//		}
-//	}
-//
-//	txsPercentage := 95 // flush cache at >95% capacity
-//	blocksProcessed := 0
-//	for {
-//		log.Debugf("indexing txs: %v", hh)
-//
-//		bh, b, err := s.headerAndBlock(ctx, hh.Hash)
-//		if err != nil {
-//			return 0, last, err
-//		}
-//
-//		// Index block
-//		err = processTxs(b.Hash(), b.Transactions(), txs)
-//		if err != nil {
-//			return 0, last, fmt.Errorf("process txs %v: %w", hh, err)
-//		}
-//
-//		blocksProcessed++
-//
-//		// Try not to overshoot the cache to prevent costly allocations
-//		cp := len(txs) * 100 / s.cfg.MaxCachedTxs
-//		if bh.Height%10000 == 0 || cp > txsPercentage || blocksProcessed == 1 {
-//			log.Infof("Tx indexer: %v tx cache %v%%", hh, cp)
-//		}
-//		if cp > txsPercentage {
-//			// Set txsMax to the largest tx capacity seen
-//			s.cfg.MaxCachedTxs = max(len(txs), s.cfg.MaxCachedTxs)
-//			last = hh
-//			// Flush
-//			break
-//		}
-//
-//		// Exit if we processed the provided end hash
-//		if endHash.IsEqual(&hh.Hash) {
-//			last = hh
-//			break
-//		}
-//
-//		// Move to next block
-//		hh, err = s.nextCanonicalBlockheader(ctx, endHash, hh)
-//		if err != nil {
-//			return 0, last, fmt.Errorf("tx next block %v: %w", hh, err)
-//		}
-//	}
-//
-//	return blocksProcessed, last, nil
-//}
-//
-//// unindexTxsInBlocks indexes txs from the last processed block until the
-//// provided end hash, inclusive. It returns the number of blocks processed and
-//// the last hash it has processed.
-//func (s *Server) unindexTxsInBlocks(ctx context.Context, endHash *chainhash.Hash, txs map[tbcd.TxKey]*tbcd.TxValue) (int, *HashHeight, error) {
-//	log.Tracef("unindexTxsInBlocks")
-//	defer log.Tracef("unindexTxsInBlocks exit")
-//
-//	// indicates if we have processed endHash and thus have hit the exit
-//	// condition.
-//	var last *HashHeight
-//
-//	// Find start hash
-//	txHH, err := s.TxIndexHash(ctx)
-//	if err != nil {
-//		return 0, last, fmt.Errorf("tx index hash: %w", err)
-//	}
-//
-//	txsPercentage := 95 // flush cache at >95% capacity
-//	blocksProcessed := 0
-//	hh := txHH
-//	for {
-//		log.Debugf("unindexing txs: %v", hh)
-//
-//		hash := hh.Hash
-//
-//		// Exit if we processed the provided end hash
-//		if endHash.IsEqual(&hash) {
-//			last = hh
-//			break
-//		}
-//
-//		bh, err := s.db.BlockHeaderByHash(ctx, hash)
-//		if err != nil {
-//			return 0, last, fmt.Errorf("block header %v: %w", hash, err)
-//		}
-//
-//		// Index block
-//		b, err := s.db.BlockByHash(ctx, bh.Hash)
-//		if err != nil {
-//			return 0, last, fmt.Errorf("block by hash %v: %w", bh, err)
-//		}
-//
-//		err = processTxs(b.Hash(), b.Transactions(), txs)
-//		if err != nil {
-//			return 0, last, fmt.Errorf("process txs %v: %w", hh, err)
-//		}
-//
-//		// This is probably not needed here since we already dealt with
-//		// it via the utxo unindexer but since it will be mostly a
-//		// no-op just go ahead.
-//		// if s.cfg.MempoolEnabled {
-//		//	// XXX this may not be the right spot.
-//		//	txHashes, _ := b.MsgBlock().TxHashes()
-//		//	_ = s.mempool.txsRemove(ctx, txHashes)
-//		// }
-//
-//		blocksProcessed++
-//
-//		// Try not to overshoot the cache to prevent costly allocations
-//		cp := len(txs) * 100 / s.cfg.MaxCachedTxs
-//		if bh.Height%10000 == 0 || cp > txsPercentage || blocksProcessed == 1 {
-//			log.Infof("Tx unindexer: %v tx cache %v%%", hh, cp)
-//		}
-//
-//		// Move to previous block
-//		height := bh.Height - 1
-//		pbh, err := s.db.BlockHeaderByHash(ctx, *bh.ParentHash())
-//		if err != nil {
-//			return 0, last, fmt.Errorf("block headers by height %v: %w",
-//				height, err)
-//		}
-//		hh.Hash = *pbh.BlockHash()
-//		hh.Height = pbh.Height
-//
-//		// We check overflow AFTER obtaining the previous hash so that
-//		// we can update the database with the LAST processed block.
-//		if cp > txsPercentage {
-//			// Set txsMax to the largest tx capacity seen
-//			s.cfg.MaxCachedTxs = max(len(txs), s.cfg.MaxCachedTxs)
-//			last = hh
-//			// Flush
-//			break
-//		}
-//	}
-//
-//	return blocksProcessed, last, nil
-//}
-//
-//func (s *Server) txIndexerUnwind(ctx context.Context, startBH, endBH *tbcd.BlockHeader) error {
-//	log.Tracef("txIndexerUnwind")
-//	defer log.Tracef("txIndexerUnwind exit")
-//
-//	// XXX dedup with txIndexerWind; it's basically the same code but with the direction, start anf endhas flipped
-//
-//	s.mtx.Lock()
-//	if !s.indexing {
-//		// XXX this prob should be an error but pusnish bad callers for now
-//		s.mtx.Unlock()
-//		panic("txIndexerUnwind indexing not true")
-//	}
-//	s.mtx.Unlock()
-//	// Allocate here so that we don't waste space when not indexing.
-//	txs := make(map[tbcd.TxKey]*tbcd.TxValue, s.cfg.MaxCachedTxs)
-//	defer clear(txs)
-//
-//	log.Infof("Start unwinding Txs at hash %v height %v", startBH, startBH.Height)
-//	log.Infof("End unwinding Txs at hash %v height %v", endBH, endBH.Height)
-//	endHash := endBH.BlockHash()
-//	for {
-//		start := time.Now()
-//		blocksProcessed, last, err := s.unindexTxsInBlocks(ctx, endHash, txs)
-//		if err != nil {
-//			return fmt.Errorf("unindex txs in blocks: %w", err)
-//		}
-//		if blocksProcessed == 0 {
-//			return nil
-//		}
-//		txsCached := len(txs)
-//		log.Infof("Tx unwinder blocks processed %v in %v transactions cached %v cache unused %v avg tx/blk %v",
-//			blocksProcessed, time.Since(start), txsCached,
-//			s.cfg.MaxCachedTxs-txsCached, txsCached/blocksProcessed)
-//
-//		// Flush to disk
-//		start = time.Now()
-//		if err = s.db.BlockTxUpdate(ctx, -1, txs, last.Hash); err != nil {
-//			return fmt.Errorf("block tx update: %w", err)
-//		}
-//		// leveldb does all kinds of allocations, force GC to lower
-//		// memory pressure.
-//		logMemStats()
-//		runtime.GC()
-//
-//		log.Infof("Flushing unwind txs complete %v took %v",
-//			txsCached, time.Since(start))
-//
-//		if endHash.IsEqual(&last.Hash) {
-//			break
-//		}
-//	}
-//	return nil
-//}
-//
-//func (s *Server) txIndexerWind(ctx context.Context, startBH, endBH *tbcd.BlockHeader) error {
-//	log.Tracef("txIndexerWind")
-//	defer log.Tracef("txIndexerWind exit")
-//
-//	s.mtx.Lock()
-//	if !s.indexing {
-//		// XXX this prob should be an error but pusnish bad callers for now
-//		s.mtx.Unlock()
-//		panic("txIndexerWind not true")
-//	}
-//	s.mtx.Unlock()
-//
-//	// Allocate here so that we don't waste space when not indexing.
-//	txs := make(map[tbcd.TxKey]*tbcd.TxValue, s.cfg.MaxCachedTxs)
-//	defer clear(txs)
-//
-//	log.Infof("Start indexing Txs at hash %v height %v", startBH, startBH.Height)
-//	log.Infof("End indexing Txs at hash %v height %v", endBH, endBH.Height)
-//	endHash := endBH.BlockHash()
-//	for {
-//		start := time.Now()
-//		blocksProcessed, last, err := s.indexTxsInBlocks(ctx, endHash, txs)
-//		if err != nil {
-//			return fmt.Errorf("index blocks: %w", err)
-//		}
-//		if blocksProcessed == 0 {
-//			return nil
-//		}
-//		txsCached := len(txs)
-//		log.Infof("Tx indexer blocks processed %v in %v transactions cached %v cache unused %v avg tx/blk %v",
-//			blocksProcessed, time.Since(start), txsCached,
-//			s.cfg.MaxCachedTxs-txsCached, txsCached/blocksProcessed)
-//
-//		// Flush to disk
-//		start = time.Now()
-//		if err = s.db.BlockTxUpdate(ctx, 1, txs, last.Hash); err != nil {
-//			return fmt.Errorf("block tx update: %w", err)
-//		}
-//		// leveldb does all kinds of allocations, force GC to lower
-//		// memory pressure.
-//		logMemStats()
-//		runtime.GC()
-//
-//		log.Infof("Flushing txs complete %v took %v",
-//			txsCached, time.Since(start))
-//
-//		if endHash.IsEqual(&last.Hash) {
-//			break
-//		}
-//	}
-//
-//	return nil
-//}
-//
-//func (s *Server) txIndexer(ctx context.Context, endHash chainhash.Hash) error {
-//	log.Tracef("txIndexer")
-//	defer log.Tracef("txIndexer exit")
-//
-//	// XXX this is basically duplicate from TxIndexIsLinear
-//
-//	s.mtx.Lock()
-//	if !s.indexing {
-//		// XXX this prob should be an error but pusnish bad callers for now
-//		s.mtx.Unlock()
-//		panic("txIndexer not true")
-//	}
-//	s.mtx.Unlock()
-//
-//	// Verify exit condition hash
-//	endBH, err := s.db.BlockHeaderByHash(ctx, endHash)
-//	if err != nil {
-//		return fmt.Errorf("blockheader hash: %w", err)
-//	}
-//
-//	// Verify start point is not after the end point
-//	txHH, err := s.TxIndexHash(ctx)
-//	if err != nil {
-//		return fmt.Errorf("tx index hash: %w", err)
-//	}
-//
-//	// Make sure there is no gap between start and end or vice versa.
-//	startBH, err := s.db.BlockHeaderByHash(ctx, txHH.Hash)
-//	if err != nil {
-//		return fmt.Errorf("blockheader hash: %w", err)
-//	}
-//	direction, err := s.TxIndexIsLinear(ctx, endHash)
-//	if err != nil {
-//		return fmt.Errorf("tx index is linear: %w", err)
-//	}
-//	switch direction {
-//	case 1:
-//		return s.txIndexerWind(ctx, startBH, endBH)
-//	case -1:
-//		return s.txIndexerUnwind(ctx, startBH, endBH)
-//	case 0:
-//		// Because we call TxIndexIsLinear we know it's the same block.
-//		return nil
-//	}
-//
-//	return fmt.Errorf("invalid direction: %v", direction)
-//}
 
 func processKeystones(block *btcutil.Block, direction int, kssCache map[chainhash.Hash]tbcd.Keystone) error {
 	if block.Height() == btcutil.BlockHeightUnknown {
@@ -1481,329 +824,6 @@ func processKeystones(block *btcutil.Block, direction int, kssCache map[chainhas
 	return nil
 }
 
-//// indexKeystonesInBlocks indexes txs from the last processed block until the
-//// provided end hash, inclusive. It returns the number of blocks processed and
-//// the last hash it processed.
-//func (s *Server) indexKeystonesInBlocks(ctx context.Context, endHash *chainhash.Hash, kss map[chainhash.Hash]tbcd.Keystone) (int, *HashHeight, error) {
-//	log.Tracef("indexKeystonesInBlocks")
-//	defer log.Tracef("indexKeystonesInBlocks exit")
-//
-//	// indicates if we have processed endHash and thus have hit the exit
-//	// condition.
-//	var last *HashHeight
-//
-//	// Find start hash
-//	keystoneHH, err := s.KeystoneIndexHash(ctx)
-//	if err != nil {
-//		return 0, last, fmt.Errorf("keystone index hash: %w", err)
-//	}
-//
-//	// If we have a real block move forward to the next block since we
-//	// already indexed the last block.
-//	hh := keystoneHH
-//	if !hh.Hash.IsEqual(s.chainParams.GenesisHash) {
-//		hh, err = s.nextCanonicalBlockheader(ctx, endHash, hh)
-//		if err != nil {
-//			return 0, last, fmt.Errorf("keystone next block %v: %w", hh, err)
-//		}
-//	} else {
-//		// Keystone indexer is special, override genesis.
-//		hh = s.hemiGenesis
-//	}
-//
-//	keystonesPercentage := 95 // flush cache at >95% capacity
-//	blocksProcessed := 0
-//	for {
-//		log.Debugf("indexing keystones: %v", hh)
-//
-//		bh, b, err := s.headerAndBlock(ctx, hh.Hash)
-//		if err != nil {
-//			return 0, last, err
-//		}
-//
-//		// Index block
-//		err = processKeystones(b, 1, kss)
-//		if err != nil {
-//			return 0, last, fmt.Errorf("process keystones %v: %w", hh, err)
-//		}
-//
-//		blocksProcessed++
-//
-//		// Try not to overshoot the cache to prevent costly allocations
-//		cp := len(kss) * 100 / s.cfg.MaxCachedKeystones
-//		if bh.Height%10000 == 0 || cp > keystonesPercentage || blocksProcessed == 1 {
-//			log.Infof("Keystone indexer: %v keystone cache %v%%", hh, cp)
-//		}
-//
-//		if cp > keystonesPercentage {
-//			// Set keystonesMax to the largest keystone capacity seen
-//			s.cfg.MaxCachedKeystones = max(len(kss), s.cfg.MaxCachedKeystones)
-//			last = hh
-//			// Flush
-//			break
-//		}
-//
-//		// Exit if we processed the provided end hash
-//		if endHash.IsEqual(&hh.Hash) {
-//			last = hh
-//			break
-//		}
-//
-//		// Move to next block
-//		hh, err = s.nextCanonicalBlockheader(ctx, endHash, hh)
-//		if err != nil {
-//			return 0, last, fmt.Errorf("keystone next block %v: %w", hh, err)
-//		}
-//	}
-//
-//	return blocksProcessed, last, nil
-//}
-//
-//// unindexKeystonesInBlocks unindexes keystones from the last processed block
-//// until the provided end hash, inclusive. It returns the number of blocks
-//// processed and the last hash it processed.
-//func (s *Server) unindexKeystonesInBlocks(ctx context.Context, endHash *chainhash.Hash, kss map[chainhash.Hash]tbcd.Keystone) (int, *HashHeight, error) {
-//	log.Tracef("unindexKeystonesInBlocks")
-//	defer log.Tracef("unindexKeystonesInBlocks exit")
-//
-//	// indicates if we have processed endHash and thus have hit the exit
-//	// condition.
-//	var last *HashHeight
-//
-//	// Find start hash
-//	ksHH, err := s.KeystoneIndexHash(ctx)
-//	if err != nil {
-//		return 0, last, fmt.Errorf("keystone index hash: %w", err)
-//	}
-//
-//	kssPercentage := 95 // flush cache at >95% capacity
-//	blocksProcessed := 0
-//	hh := ksHH
-//	for {
-//		log.Debugf("unindexing keystones: %v", hh)
-//
-//		hash := hh.Hash
-//
-//		// Exit if we processed the provided end hash
-//		if endHash.IsEqual(&hash) {
-//			last = hh
-//			break
-//		}
-//
-//		bh, b, err := s.headerAndBlock(ctx, hh.Hash)
-//		if err != nil {
-//			return 0, last, err
-//		}
-//
-//		err = processKeystones(b, -1, kss)
-//		if err != nil {
-//			return 0, last, fmt.Errorf("process keystones %v: %w", hh, err)
-//		}
-//
-//		blocksProcessed++
-//
-//		// Try not to overshoot the cache to prevent costly allocations
-//		cp := len(kss) * 100 / s.cfg.MaxCachedKeystones
-//		if bh.Height%10000 == 0 || cp > kssPercentage || blocksProcessed == 1 {
-//			log.Infof("Keystone unindexer: %v keystone cache %v%%", hh, cp)
-//		}
-//
-//		// Move to previous block
-//		height := bh.Height - 1
-//		pbh, err := s.db.BlockHeaderByHash(ctx, *bh.ParentHash())
-//		if err != nil {
-//			return 0, last, fmt.Errorf("block headers by height %v: %w",
-//				height, err)
-//		}
-//		hh.Hash = *pbh.BlockHash()
-//		hh.Height = pbh.Height
-//
-//		// We check overflow AFTER obtaining the previous hash so that
-//		// we can update the database with the LAST processed block.
-//		if cp > kssPercentage {
-//			// Set kssMax to the largest keystone capacity seen
-//			s.cfg.MaxCachedKeystones = max(len(kss), s.cfg.MaxCachedKeystones)
-//			last = hh
-//			// Flush
-//			break
-//		}
-//	}
-//
-//	return blocksProcessed, last, nil
-//}
-//
-//func (s *Server) keystoneIndexerUnwind(ctx context.Context, startBH, endBH *tbcd.BlockHeader) error {
-//	log.Tracef("keystoneIndexerUnwind")
-//	defer log.Tracef("keystoneIndexerUnwind exit")
-//
-//	// XXX dedup with keystoneIndexerWind; it's basically the same code but with the direction, start anf endhas flipped
-//
-//	s.mtx.Lock()
-//	if !s.indexing {
-//		// XXX this prob should be an error but pusnish bad callers for now
-//		s.mtx.Unlock()
-//		panic("keystoneIndexerUnwind indexing not true")
-//	}
-//	s.mtx.Unlock()
-//	// Allocate here so that we don't waste space when not indexing.
-//	kss := make(map[chainhash.Hash]tbcd.Keystone, s.cfg.MaxCachedKeystones)
-//	defer clear(kss)
-//
-//	log.Infof("Start unwinding keystones at hash %v height %v", startBH, startBH.Height)
-//	log.Infof("End unwinding keystones at hash %v height %v", endBH, endBH.Height)
-//	endHash := endBH.BlockHash()
-//	for {
-//		start := time.Now()
-//		blocksProcessed, last, err := s.unindexKeystonesInBlocks(ctx, endHash, kss)
-//		if err != nil {
-//			return fmt.Errorf("unindex keystones in blocks: %w", err)
-//		}
-//		if blocksProcessed == 0 {
-//			return nil
-//		}
-//		kssCached := len(kss)
-//		log.Infof("Keystone unwinder blocks processed %v in %v keystones cached %v cache unused %v avg keystone/blk %v",
-//			blocksProcessed, time.Since(start), kssCached,
-//			s.cfg.MaxCachedKeystones-kssCached, kssCached/blocksProcessed)
-//
-//		// Flush to disk
-//		start = time.Now()
-//		if err = s.db.BlockKeystoneUpdate(ctx, -1, kss, last.Hash); err != nil {
-//			return fmt.Errorf("block keystone update: %w", err)
-//		}
-//		// leveldb does all kinds of allocations, force GC to lower
-//		// memory pressure.
-//		logMemStats()
-//		runtime.GC()
-//
-//		log.Infof("Flushing unwind keystones complete %v took %v",
-//			kssCached, time.Since(start))
-//
-//		if endHash.IsEqual(&last.Hash) {
-//			break
-//		}
-//	}
-//	return nil
-//}
-//
-//func (s *Server) keystoneIndexerWind(ctx context.Context, startBH, endBH *tbcd.BlockHeader) error {
-//	log.Tracef("keystoneIndexerWind")
-//	defer log.Tracef("keystoneIndexerWind exit")
-//
-//	s.mtx.Lock()
-//	if !s.indexing {
-//		// XXX this prob should be an error but pusnish bad callers for now
-//		s.mtx.Unlock()
-//		panic("keystoneIndexerWind not true")
-//	}
-//	s.mtx.Unlock()
-//
-//	// Allocate here so that we don't waste space when not indexing.
-//	kss := make(map[chainhash.Hash]tbcd.Keystone, s.cfg.MaxCachedKeystones)
-//	defer clear(kss)
-//
-//	log.Infof("Start indexing keystones at hash %v height %v", startBH, startBH.Height)
-//	log.Infof("End indexing keystones at hash %v height %v", endBH, endBH.Height)
-//	endHash := endBH.BlockHash()
-//	for {
-//		start := time.Now()
-//		blocksProcessed, last, err := s.indexKeystonesInBlocks(ctx, endHash, kss)
-//		if err != nil {
-//			return fmt.Errorf("index blocks: %w", err)
-//		}
-//		if blocksProcessed == 0 {
-//			return nil
-//		}
-//		kssCached := len(kss)
-//		log.Infof("Keystone indexer blocks processed %v in %v keystones cached %v cache unused %v avg keystones/blk %v",
-//			blocksProcessed, time.Since(start), kssCached,
-//			s.cfg.MaxCachedKeystones-kssCached, kssCached/blocksProcessed)
-//
-//		// Flush to disk
-//		start = time.Now()
-//		if err = s.db.BlockKeystoneUpdate(ctx, 1, kss, last.Hash); err != nil {
-//			return fmt.Errorf("block hemi update: %w", err)
-//		}
-//		// leveldb does all kinds of allocations, force GC to lower
-//		// memory pressure.
-//		logMemStats()
-//		runtime.GC()
-//
-//		log.Infof("Flushing keystones complete %v took %v",
-//			kssCached, time.Since(start))
-//
-//		if endHash.IsEqual(&last.Hash) {
-//			break
-//		}
-//	}
-//
-//	return nil
-//}
-//
-//func (s *Server) keystoneIndexer(ctx context.Context, endHash chainhash.Hash) error {
-//	log.Tracef("keystoneIndexer")
-//	defer log.Tracef("keystoneIndexer exit")
-//
-//	// XXX this is basically duplicate from KeystoneIndexIsLinear
-//
-//	if !s.cfg.HemiIndex {
-//		return errors.New("disabled")
-//	}
-//	s.mtx.Lock()
-//	if !s.indexing {
-//		// XXX this prob should be an error but pusnish bad callers for now
-//		s.mtx.Unlock()
-//		panic("keystoneIndexer not true")
-//	}
-//	s.mtx.Unlock()
-//
-//	// Verify exit condition hash
-//	endBH, err := s.db.BlockHeaderByHash(ctx, endHash)
-//	if err != nil {
-//		return fmt.Errorf("blockheader end hash: %w", err)
-//	}
-//
-//	// Verify start point is not after the end point
-//	keystoneHH, err := s.KeystoneIndexHash(ctx)
-//	if err != nil {
-//		return fmt.Errorf("keystone index hash: %w", err)
-//	}
-//
-//	// Make sure there is no gap between start and end or vice versa.
-//	startBH, err := s.db.BlockHeaderByHash(ctx, keystoneHH.Hash)
-//	if err != nil {
-//		return fmt.Errorf("blockheader keystone hash: %w", err)
-//	}
-//	direction, err := s.KeystoneIndexIsLinear(ctx, endHash)
-//	if err != nil {
-//		return fmt.Errorf("keystone index is linear: %w", err)
-//	}
-//	switch direction {
-//	case 1:
-//		return s.keystoneIndexerWind(ctx, startBH, endBH)
-//	case -1:
-//		return s.keystoneIndexerUnwind(ctx, startBH, endBH)
-//	case 0:
-//		// Because we call KeystoneIndexIsLinear we know it's the same block.
-//		return nil
-//	}
-//
-//	return fmt.Errorf("invalid direction: %v", direction)
-//}
-
-//	func (s *Server) UtxoIndexIsLinear(ctx context.Context, endHash chainhash.Hash) (int, error) {
-//		log.Tracef("UtxoIndexIsLinear")
-//		defer log.Tracef("UtxoIndexIsLinear exit")
-//
-//		// Verify start point is not after the end point
-//		utxoHH, err := s.UtxoIndexHash(ctx)
-//		if err != nil {
-//			return 0, fmt.Errorf("utxo index hash: %w", err)
-//		}
-//
-//		return s.IndexIsLinear(ctx, utxoHH.Hash, endHash)
-//	}
-
 // TxIndexIsLinear do we really want this exported?
 func (s *Server) TxIndexIsLinear(ctx context.Context, endHash chainhash.Hash) (int, error) {
 	log.Tracef("TxIndexIsLinear")
@@ -1817,19 +837,6 @@ func (s *Server) TxIndexIsLinear(ctx context.Context, endHash chainhash.Hash) (i
 
 	return s.IndexIsLinear(ctx, txHH.Hash, endHash)
 }
-
-//func (s *Server) KeystoneIndexIsLinear(ctx context.Context, endHash chainhash.Hash) (int, error) {
-//	log.Tracef("KeystoneIndexIsLinear")
-//	defer log.Tracef("KeystoneIndexIsLinear exit")
-//
-//	// Verify start point is not after the end point
-//	keystoneHH, err := s.KeystoneIndexHash(ctx)
-//	if err != nil {
-//		return 0, fmt.Errorf("keystone index hash: %w", err)
-//	}
-//
-//	return s.IndexIsLinear(ctx, keystoneHH.Hash, endHash)
-//}
 
 func (s *Server) IndexIsLinear(ctx context.Context, startHash, endHash chainhash.Hash) (int, error) {
 	log.Tracef("IndexIsLinear")
@@ -1951,104 +958,6 @@ func (s *Server) SyncIndexersToHash(ctx context.Context, hash chainhash.Hash) er
 	return nil
 }
 
-func (s *Server) utxoIndexersToBest(ctx context.Context, bhb *tbcd.BlockHeader) error {
-	log.Tracef("utxoIndexersToBest")
-	defer log.Tracef("utxoIndexersToBest exit")
-
-	return s.newUtxoIndexer().modeIndexersToBest(ctx, bhb)
-	//// Index Utxos to best
-	//utxoHH, err := s.UtxoIndexHash(ctx)
-	//if err != nil {
-	//	return fmt.Errorf("utxo index hash: %w", err)
-	//}
-	//utxoBH, err := s.db.BlockHeaderByHash(ctx, utxoHH.Hash)
-	//if err != nil {
-	//	return err
-	//}
-	//cp, err := s.findCanonicalParent(ctx, utxoBH)
-	//if err != nil {
-	//	return err
-	//}
-	//if !cp.Hash.IsEqual(&utxoBH.Hash) {
-	//	log.Infof("Syncing utxo index to: %v from: %v via: %v",
-	//		bhb.HH(), utxoBH.HH(), cp.HH())
-	//	// utxoBH is NOT on canonical chain, unwind first
-	//	if err := s.utxoIndexer(ctx, cp.Hash); err != nil {
-	//		return fmt.Errorf("utxo indexer unwind: %w", err)
-	//	}
-	//}
-	//// Index utxo to best block
-	//if err := s.utxoIndexer(ctx, bhb.Hash); err != nil {
-	//	return fmt.Errorf("utxo indexer: %w", err)
-	//}
-
-	// return nil
-}
-
-func (s *Server) txIndexersToBest(ctx context.Context, bhb *tbcd.BlockHeader) error {
-	log.Tracef("txIndexersToBest")
-	defer log.Tracef("txIndexersToBest exit")
-
-	return s.newTxIndexer().modeIndexersToBest(ctx, bhb)
-	//// Index Tx
-	//txHH, err := s.TxIndexHash(ctx)
-	//if err != nil {
-	//	return fmt.Errorf("tx index hash: %w", err)
-	//}
-	//txBH, err := s.db.BlockHeaderByHash(ctx, txHH.Hash)
-	//if err != nil {
-	//	return err
-	//}
-	//cp, err := s.findCanonicalParent(ctx, txBH)
-	//if err != nil {
-	//	return err
-	//}
-	//if !cp.Hash.IsEqual(&txBH.Hash) {
-	//	log.Infof("Syncing tx index to: %v from: %v via: %v",
-	//		bhb.HH(), txBH.HH(), cp.HH())
-	//	// txBH is NOT on canonical chain, unwind first
-	//	if err := s.txIndexer(ctx, cp.Hash); err != nil {
-	//		return fmt.Errorf("tx indexer unwind: %w", err)
-	//	}
-	//}
-	//// Transactions index
-	//if err := s.txIndexer(ctx, bhb.Hash); err != nil {
-	//	return fmt.Errorf("tx indexer: %w", err)
-	//}
-}
-
-func (s *Server) keystoneIndexersToBest(ctx context.Context, bhb *tbcd.BlockHeader) error {
-	log.Tracef("keystoneIndexersToBest")
-	defer log.Tracef("keystoneIndexersToBest exit")
-
-	return s.newKeystoneIndexer().modeIndexersToBest(ctx, bhb)
-	//// Index keystones to best
-	//keystoneHH, err := s.KeystoneIndexHash(ctx)
-	//if err != nil {
-	//	return fmt.Errorf("keystone index hash: %w", err)
-	//}
-	//keystoneBH, err := s.db.BlockHeaderByHash(ctx, keystoneHH.Hash)
-	//if err != nil {
-	//	return err
-	//}
-	//cp, err := s.findCanonicalParent(ctx, keystoneBH)
-	//if err != nil {
-	//	return err
-	//}
-	//if !cp.Hash.IsEqual(&keystoneBH.Hash) {
-	//	log.Infof("Syncing keystone index to: %v from: %v via: %v",
-	//		bhb.HH(), keystoneBH.HH(), cp.HH())
-	//	// keystoneBH is NOT on canonical chain, unwind first
-	//	if err := s.keystoneIndexer(ctx, cp.Hash); err != nil {
-	//		return fmt.Errorf("keystone indexer unwind: %w", err)
-	//	}
-	//}
-	//// Index keystones to best block
-	//if err := s.keystoneIndexer(ctx, bhb.Hash); err != nil {
-	//	return fmt.Errorf("keystone indexer: %w", err)
-	//}
-}
-
 func (s *Server) syncIndexersToBest(ctx context.Context) error {
 	log.Tracef("syncIndexersToBest")
 	defer log.Tracef("syncIndexersToBest exit")
@@ -2060,16 +969,16 @@ func (s *Server) syncIndexersToBest(ctx context.Context) error {
 
 	log.Debugf("Sync indexers to best: %v @ %v", bhb, bhb.Height)
 
-	if err := s.utxoIndexersToBest(ctx, bhb); err != nil {
+	if err := s.newUtxoIndexer().modeIndexersToBest(ctx, bhb); err != nil {
 		return err
 	}
 
-	if err := s.txIndexersToBest(ctx, bhb); err != nil {
+	if err := s.newTxIndexer().modeIndexersToBest(ctx, bhb); err != nil {
 		return err
 	}
 
 	if s.cfg.HemiIndex {
-		if err := s.keystoneIndexersToBest(ctx, bhb); err != nil {
+		if err := s.newKeystoneIndexer().modeIndexersToBest(ctx, bhb); err != nil {
 			return err
 		}
 	}
